fix(cli): harden GitHub identity reads and gate error fallback
Move saved_github_username into a lightweight github_identity module so the
welcome banner does not import the heavy MCP stack, restore try/except around
banner reads, and detect CI bypasses via env vars only in the gate error path.

Fixes #3133
