+++ integrations/github_login.py
@@ -34,24 +34,6 @@ class GitHubLoginResult:
     detail: str = ""
 
 
-def saved_github_username() -> str:
-    """Return the persisted GitHub login from the integration store, or "".
-
-    Best-effort and never raises: callers like the welcome banner and analytics
-    re-identify must work even when the store is unreadable.
-    """
-    try:
-        from integrations.store import get_integration
-
-        record = get_integration("github")
-        if not record:
-            return ""
-        credentials = record.get("credentials") or {}
-        return str(credentials.get("username") or "").strip()
-    except Exception:
-        return ""
-
-
 def authenticate_and_configure_github(
     *,
     on_prompt: Callable[[GitHubDeviceCode], None] | None = None,
